+++ stream-chat-android-state/src/test/java/io/getstream/chat/android/state/channel/controller/WhenHandleEvent.kt
@@ -76,7 +76,8 @@ internal class WhenHandleEvent : SynchronizedCoroutineTest {
         MutableStateFlow(
             mapOf(currentUser.id to currentUser),
         ),
-    )
+    ) { System.currentTimeMillis() }
+
     private val channelStateLogic: ChannelStateLogic = mock {
         on(it.writeChannelState()) doReturn channelMutableState
     }