+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/feature/messages/list/MessageListController.kt
@@ -412,6 +412,7 @@ public class MessageListController(
      * load the message if it is not in the list and scroll to it.
      */
     init {
+        logger.i { "<init> cid: $cid, messageId: $messageId, messageLimit: $messageLimit" }
         observeMessagesListState()
         processMessageId()
     }
@@ -579,7 +580,7 @@ public class MessageListController(
         messageId
             ?.takeUnless { it.isBlank() }
             ?.let { messageId ->
-                logger.i { "[processMessageId] messageId: $messageId, parentMessageId: $parentMessageId" }
+                logger.d { "[processMessageId] messageId: $messageId, parentMessageId: $parentMessageId" }
                 scope.launch {
                     if (parentMessageId != null) {
                         enterThreadSequential(parentMessageId)
@@ -1021,7 +1022,7 @@ public class MessageListController(
      * @param messageLimit The size of the message list page to load.
      */
     public fun loadNewerMessages(baseMessageId: String, messageLimit: Int = this.messageLimit) {
-        logger.i { "[loadNewerMessages] baseMessageId: $baseMessageId, messageLimit: $messageLimit" }
+        logger.d { "[loadNewerMessages] baseMessageId: $baseMessageId, messageLimit: $messageLimit" }
         if (clientState.isOffline) return
         _mode.value.run {
             when (this) {
@@ -1079,7 +1080,7 @@ public class MessageListController(
      * @param messageLimit The size of the message list page to load.
      */
     public fun loadOlderMessages(messageLimit: Int = this.messageLimit) {
-        logger.i { "[loadOlderMessages] messageLimit: $messageLimit" }
+        logger.d { "[loadOlderMessages] messageLimit: $messageLimit" }
         if (clientState.isOffline) return
 
         _mode.value.run {
@@ -1213,7 +1214,7 @@ public class MessageListController(
      * @param onResult Handler that notifies the result of the load action.
      */
     public fun loadMessageById(messageId: String, onResult: (Result<Message>) -> Unit = {}) {
-        logger.i { "[loadMessageById] messageId: $messageId" }
+        logger.d { "[loadMessageById] messageId: $messageId" }
         chatClient.loadMessageById(cid, messageId).enqueue { result ->
             onResult(result)
             if (result is Result.Failure) {
@@ -1550,7 +1551,7 @@ public class MessageListController(
 
         val lastSeenMessageId = this.lastSeenMessageId
         if (lastSeenMessageId == messageId) {
-            logger.w { "[markLastMessageRead] cid: $cid; rejected[$isInThread] (already seen msgId): $messageId" }
+            logger.v { "[markLastMessageRead] cid: $cid; rejected[$isInThread] (already seen msgId): $messageId" }
             return
         }
         this.lastSeenMessageId = messageId