+++ stream-chat-android-ui-uitests/src/androidTest/java/io/getstream/chat/android/uitests/snapshot/compose/messages/MessageItemTest.kt
@@ -17,6 +17,7 @@
 package io.getstream.chat.android.uitests.snapshot.compose.messages
 
 import io.getstream.chat.android.compose.ui.messages.list.MessageItem
+import io.getstream.chat.android.models.ReactionSortingByCount
 import io.getstream.chat.android.ui.common.state.messages.list.MessageItemState
 import io.getstream.chat.android.uitests.snapshot.compose.ComposeScreenshotTest
 import io.getstream.chat.android.uitests.util.TestData
@@ -33,6 +34,7 @@ class MessageItemTest : ComposeScreenshotTest() {
                 showMessageFooter = true,
             ),
             onLongItemClick = {},
+            reactionSorting = ReactionSortingByCount,
         )
     }
 
@@ -45,6 +47,7 @@ class MessageItemTest : ComposeScreenshotTest() {
                 showMessageFooter = true,
             ),
             onLongItemClick = {},
+            reactionSorting = ReactionSortingByCount,
         )
     }
 
@@ -58,6 +61,7 @@ class MessageItemTest : ComposeScreenshotTest() {
                 showMessageFooter = true,
             ),
             onLongItemClick = {},
+            reactionSorting = ReactionSortingByCount,
         )
     }
 }